Styled up home page, summary header links now link to their respective examples. Reformatted navigation at the top. Also, renamed some of the example components to be more semantic and fixed the folder placement.
